handling rb2 - scheduler

+++ hoversprite-principle/hoversprite-main-component/hoversprite-spray-component/src/main/java/hoversprite/project/sprayerAssignment/SprayerAssignmentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface SprayerAssignmentRepository extends JpaRepository<SprayerAssignment, Integer>, SprayerAssignmentRepositoryCustom {
+interface SprayerAssignmentRepository extends JpaRepository<SprayerAssignment, Long>, SprayerAssignmentRepositoryCustom {
 }
